Translate: Fail on errors in structure of translations

+++ trubar/tests/test_command_line.sh
@@ -68,8 +68,20 @@ then
   exit 1
 fi
 
-print_run 'trubar translate -s test_project test_project/translations.yaml -q -n'
-diff -r si_translations si_translations_copy
+echo "... faulty translations"
+print_run 'trubar translate -s test_project -d test_translations test_project/faulty_translations.yaml -q'
+diff -r si_translations/submodule/apples.py test_translations/submodule/apples.py
+if [ -f test_translations/__init__.py ]
+then
+  echo "Wrote errored file."
+  exit 1
+fi
+if [ -f test_translations/trash/nothing.py ]
+then
+  echo "Wrote errored file."
+  exit 1
+fi
+rm -r test_translations
 rm -r si_translations_copy
 
 echo
